services/ai-engine/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.34.0
pypdf2==3.0.1
python-docx==1.1.2
python-multipart==0.0.9
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
tavily-python==0.4.0
